+++ backend/src/main/java/com/pickeat/backend/room/ui/RoomController.java
@@ -2,7 +2,7 @@ package com.pickeat.backend.room.ui;
 
 import com.pickeat.backend.global.auth.annotation.LoginUserId;
 import com.pickeat.backend.global.log.BusinessLogging;
-import com.pickeat.backend.room.application.dto.RoomService;
+import com.pickeat.backend.room.application.RoomService;
 import com.pickeat.backend.room.application.dto.request.RoomInvitationRequest;
 import com.pickeat.backend.room.application.dto.request.RoomRequest;
 import com.pickeat.backend.room.application.dto.response.RoomResponse;
